requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.40.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.5.0
huggingface-hub>=0.16.0
accelerate>=0.20.0
python-dotenv>=1.0.0
